Adds the first part of the Book API
