+++ app/src/main/java/com/github/sdp/mediato/HomeViewModel.java
@@ -3,12 +3,10 @@ package com.github.sdp.mediato;
 
 import android.app.Application;
 
-import androidx.annotation.Nullable;
 import androidx.lifecycle.AndroidViewModel;
-import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
-import com.github.sdp.mediato.api.themoviedb.TheMovieDB;
+import com.github.sdp.mediato.api.themoviedb.TheMovieDBAPI;
 import com.github.sdp.mediato.model.media.Media;
 import com.github.sdp.mediato.model.media.Movie;
 
@@ -19,12 +17,12 @@ public class HomeViewModel  extends AndroidViewModel {
 
     Application application;
     private final MutableLiveData<List<Media>> movies = new MutableLiveData<>();
-    private final TheMovieDB api;
+    private final TheMovieDBAPI api;
 
     public HomeViewModel(Application application){
         super(application);
         this.application = application;
-        api = new TheMovieDB(application.getApplicationContext().getString(R.string.tmdb_url),
+        api = new TheMovieDBAPI(application.getApplicationContext().getString(R.string.tmdb_url),
                 application.getApplicationContext().getString(R.string.TMDBAPIKEY));
     }
     public MutableLiveData<List<Media>> getMovies() {
